+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.0.0
 - Seal `BitDecodeExt`, `BitEncodeExt`, `BitCodec` traits
+- Add `#[codec(other)]` attribute
 # v0.11.1
 - Fix potential memory leak in `BitDecode` implementation for `[T; N]`
 # v0.11.0